Update wwwverifier to Recent CD Consultation
Modified build scripts to make wwwverifier a subproject of the
Identity Credential project to reduce copy-pasted code. Updated
OriginInfo to match 18013-7 recent CD consultation.

Tested appholder changes manually with unattended presentations
via the wwwverifier as well as attended presentations with
appverifier.
