+++ .gitignore
@@ -9,3 +9,4 @@
 fastlane/report.xml
 /mdl-ref-server/mdl-server-db.sqlite3
 /version*
+wwwverifier/build/*
