+++ identity/src/test/java/com/android/identity/mdoc/origininfo/OriginInfoTest.java
@@ -26,80 +26,26 @@ import org.junit.runners.JUnit4;
 @RunWith(JUnit4.class)
 public class OriginInfoTest {
     @Test
-    public void testOriginInfoQr() {
-        OriginInfoQr info = new OriginInfoQr(OriginInfo.CAT_RECEIVE);
-        OriginInfoQr decoded = OriginInfoQr.decode(info.encode());
-        Assert.assertEquals(OriginInfo.CAT_RECEIVE, decoded.getCat());
-        Assert.assertEquals("{\n" +
-                "  'cat' : 1,\n" +
-                "  'type' : 2,\n" +
-                "  'Details' : null\n" +
-                "}", Util.cborPrettyPrint(info.encode()));
-    }
-
-    @Test
-    public void testOriginInfoQrDelivery() {
-        OriginInfoQr info = new OriginInfoQr(OriginInfo.CAT_DELIVERY);
-        OriginInfoQr decoded = OriginInfoQr.decode(info.encode());
-        Assert.assertEquals(OriginInfo.CAT_DELIVERY, decoded.getCat());
-        Assert.assertEquals("{\n" +
-                "  'cat' : 0,\n" +
-                "  'type' : 2,\n" +
-                "  'Details' : null\n" +
-                "}", Util.cborPrettyPrint(info.encode()));
-    }
-
-    @Test
-    public void testOriginInfoNfc() {
-        OriginInfoNfc info = new OriginInfoNfc(OriginInfo.CAT_RECEIVE);
-        OriginInfoNfc decoded = OriginInfoNfc.decode(info.encode());
-        Assert.assertEquals(OriginInfo.CAT_RECEIVE, decoded.getCat());
-        Assert.assertEquals("{\n" +
-                "  'cat' : 1,\n" +
-                "  'type' : 3,\n" +
-                "  'Details' : null\n" +
-                "}", Util.cborPrettyPrint(info.encode()));
-    }
-
-    @Test
-    public void testOriginInfoNfcDelivery() {
-        OriginInfoNfc info = new OriginInfoNfc(OriginInfo.CAT_DELIVERY);
-        OriginInfoNfc decoded = OriginInfoNfc.decode(info.encode());
-        Assert.assertEquals(OriginInfo.CAT_DELIVERY, decoded.getCat());
-        Assert.assertEquals("{\n" +
-                "  'cat' : 0,\n" +
-                "  'type' : 3,\n" +
-                "  'Details' : null\n" +
-                "}", Util.cborPrettyPrint(info.encode()));
-    }
-
-    @Test
-    public void testOriginInfoWebsite() {
-        OriginInfoWebsite info = new OriginInfoWebsite(OriginInfo.CAT_RECEIVE, "https://foo.com/bar");
-        OriginInfoWebsite decoded = OriginInfoWebsite.decode(info.encode());
-        Assert.assertEquals(OriginInfo.CAT_RECEIVE, decoded.getCat());
-        Assert.assertEquals("https://foo.com/bar", decoded.getBaseUrl());
+    public void testOriginInfoReferrerUrl() {
+        OriginInfoReferrerUrl info = new OriginInfoReferrerUrl("https://foo.com/bar");
+        OriginInfoReferrerUrl decoded = OriginInfoReferrerUrl.decode(info.encode());
+        Assert.assertEquals("https://foo.com/bar", decoded.getUrl());
         Assert.assertEquals("{\n" +
                 "  'cat' : 1,\n" +
                 "  'type' : 1,\n" +
-                "  'Details' : {\n" +
-                "    'baseUrl' : 'https://foo.com/bar'\n" +
-                "  }\n" +
+                "  'details' : 'https://foo.com/bar'\n" +
                 "}", Util.cborPrettyPrint(info.encode()));
     }
 
     @Test
-    public void testOriginInfoWebsiteDelivery() {
-        OriginInfoWebsite info = new OriginInfoWebsite(OriginInfo.CAT_DELIVERY, "https://foo.com/baz");
-        OriginInfoWebsite decoded = OriginInfoWebsite.decode(info.encode());
-        Assert.assertEquals(OriginInfo.CAT_DELIVERY, decoded.getCat());
+    public void testOriginInfoWebsiteBase() {
+        OriginInfoBaseUrl info = new OriginInfoBaseUrl("https://foo.com/bar");
+        OriginInfoBaseUrl decoded = OriginInfoBaseUrl.decode(info.encode());
+        Assert.assertEquals("https://foo.com/bar", decoded.getUrl());
         Assert.assertEquals("{\n" +
-                "  'cat' : 0,\n" +
-                "  'type' : 1,\n" +
-                "  'Details' : {\n" +
-                "    'baseUrl' : 'https://foo.com/baz'\n" +
-                "  }\n" +
+                "  'cat' : 1,\n" +
+                "  'type' : 2,\n" +
+                "  'details' : 'https://foo.com/bar'\n" +
                 "}", Util.cborPrettyPrint(info.encode()));
     }
-
 }
