+++ wwwverifier/build.gradle
@@ -17,26 +17,18 @@ plugins {
   id 'maven-publish'
 }
 
-repositories {
-  jcenter()
-  mavenLocal()
-  maven {
-      url = uri('https://repo.maven.apache.org/maven2/')
-  }
-}
-
 apply plugin: 'org.gretty'
 apply plugin: 'java'
 apply plugin: 'war'
 apply plugin: 'com.google.cloud.tools.appengine'
 
 dependencies {
-  //implementation project(path: ':identity', configuration:'default')
   implementation 'com.google.appengine:appengine-api-1.0-sdk:1.9.59'
   implementation 'com.google.code.gson:gson:2.9.0'
   implementation 'co.nstant.in:cbor:0.9'
   implementation 'org.json:json:20160810'
   implementation 'org.bouncycastle:bcprov-jdk15on:1.70'
+  implementation project(path: ':identity')
   testImplementation 'junit:junit:4.12'
   testImplementation 'org.mockito:mockito-core:3.2.4'
   testImplementation 'com.google.appengine:appengine-testing:1.9.64'
@@ -55,9 +47,6 @@ appengine {
   }
 }
 
-group = 'com.google.sps'
-version = '1'
-description = 'portfolio'
 java.sourceCompatibility = JavaVersion.VERSION_1_8
 java.targetCompatibility = JavaVersion.VERSION_1_8
